Move test types to shared types
